Feat: 유저 비밀번호 필드 추가
- 올바른 비밀번호 입력시에만 유저 수정 및 삭제 가능하도록 구현

+++ src/main/java/com/example/scheduleprojectdevelop/service/UserService.java
@@ -18,8 +18,8 @@ public class UserService {
     private final UserRepository userRepository;
 
     //유저 생성
-    public UserResponseDto signUp(String username, String email) {
-        User user = new User(username, email);
+    public UserResponseDto signUp(String username, String email, String password) {
+        User user = new User(username, email, password);
         User savedUser = userRepository.save(user);
         return new UserResponseDto(savedUser.getId(), savedUser.getUsername(), savedUser.getEmail());
     }
@@ -39,16 +39,23 @@ public class UserService {
 
     //유저 수정
     @Transactional
-    public UserResponseDto updateUser(Long id, String username, String email) {
+    public UserResponseDto updateUser(Long id, String username, String email, String password) {
         User user = userRepository.findUserByIdOrElseThrow(id);
-        user.update(username, email);
-        return new UserResponseDto(id, user.getUsername(), user.getEmail());
+        if(user.getPassword().equals(password)){
+            user.update(username, email);
+            return new UserResponseDto(id, user.getUsername(), user.getEmail());
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Password is wrong.");
     }
 
     //유저 삭제
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id, String password) {
         User findUser = userRepository.findUserByIdOrElseThrow(id);
-        userRepository.delete(findUser);
-        //userRepository.deleteById(id); 로 해도 됨
+        if(findUser.getPassword().equals(password)){
+            userRepository.delete(findUser);
+            //userRepository.deleteById(id); 로 해도 됨
+            return;
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Password is wrong.");
     }
 }
